Rename notification package
