+++ core/src/main/java/life/qbic/identityaccess/application/notification/EventStore.java
@@ -1,4 +1,4 @@
-package life.qbic.apps.datamanager.events;
+package life.qbic.identityaccess.application.notification;
 
 import java.util.Set;
 import life.qbic.identityaccess.domain.events.DomainEvent;
