+++ core/src/main/java/life/qbic/identityaccess/application/notification/MessageParameters.java
@@ -1,4 +1,4 @@
-package life.qbic.identityaccess.application.notifications;
+package life.qbic.identityaccess.application.notification;
 
 import java.time.Instant;
