+++ core/src/main/java/life/qbic/identityaccess/application/notification/Notification.java
@@ -1,4 +1,4 @@
-package life.qbic.identityaccess.application.notifications;
+package life.qbic.identityaccess.application.notification;
 
 import java.time.Instant;
 import life.qbic.identityaccess.domain.events.DomainEvent;
