+++ core/src/main/java/life/qbic/identityaccess/application/user/UserRegistrationService.java
@@ -7,9 +7,9 @@ import java.util.List;
 import java.util.Optional;
 import java.util.function.Consumer;
 import life.qbic.identityaccess.application.ApplicationException;
-import life.qbic.apps.datamanager.events.EventStore;
-import life.qbic.identityaccess.application.notifications.Notification;
-import life.qbic.identityaccess.application.notifications.NotificationService;
+import life.qbic.identityaccess.application.notification.EventStore;
+import life.qbic.identityaccess.application.notification.Notification;
+import life.qbic.identityaccess.application.notification.NotificationService;
 import life.qbic.identityaccess.domain.events.DomainEventPublisher;
 import life.qbic.identityaccess.domain.events.DomainEventSubscriber;
 import life.qbic.identityaccess.domain.user.EmailAddress;
