+++ webapp/src/main/java/life/qbic/Application.java
@@ -13,8 +13,8 @@ import java.io.ObjectInputStream;
 import life.qbic.email.Email;
 import life.qbic.email.EmailService;
 import life.qbic.email.Recipient;
-import life.qbic.identityaccess.application.notifications.MessageBusInterface;
-import life.qbic.identityaccess.application.notifications.MessageSubscriber;
+import life.qbic.identityaccess.application.notification.MessageBusInterface;
+import life.qbic.identityaccess.application.notification.MessageSubscriber;
 import life.qbic.identityaccess.application.user.ConfirmEmailOutput;
 import life.qbic.identityaccess.application.user.EmailAddressConfirmation;
 import life.qbic.identityaccess.domain.events.DomainRegistry;
